Problem: initializeUsers is in client scope
Solution: moved it to server. #69
